\name{print.graphicalVAR}
\alias{print.graphicalVAR}
\alias{summary.graphicalVAR}
\title{
S3 methods for graphicalVAR objects.
}
\description{
Prints a short overview of the results of \code{\link{graphicalVAR}}
}
\usage{
\method{print}{graphicalVAR}(x, ...)
\method{summary}{graphicalVAR}(object, ...)
}
\arguments{
  \item{x}{
A \code{graphicalVAR} object
}
  \item{object}{
A \code{graphicalVAR} object
}
  \item{\dots}{
Not used.
}
}
\author{
Sacha Epskamp <mail@sachaepskamp.com>
}